IA-2876: move hooks
