+++ hat/assets/js/apps/Iaso/domains/workflows/details.tsx
@@ -26,7 +26,7 @@ import { useParams } from 'react-router-dom';
 import TopBar from '../../components/nav/TopBarComponent';
 import MESSAGES from './messages';
 
-import { useGoBack } from '../../routing/useGoBack';
+import { useGoBack } from '../../routing/hooks/useGoBack';
 import { redirectToReplace } from '../../routing/actions';
 import { baseUrls } from '../../constants/urls';
